start of port

+++ src/main/java/com/sammy/malum/visual_effects/networked/data/PositionEffectData.java
@@ -35,6 +35,6 @@ public class PositionEffectData {
     }
 
     public BlockPos getAsBlockPos() {
-        return new BlockPos(posX, posY, posZ);
+        return new BlockPos((int)posX, (int)posY, (int)posZ);
     }
 }
